AddCookiesPolicy.Factory -> CookiePolicyFactory
